move scanner interface to an image-inspector type that all scanners should implement
